File's path searching
Put the search.exe into the path you want to search, input the 'sign'-which means if sign is in the file's name, the path of it would be shown in the two textbox below. Multiple signs input is available, merely to input like 'sign1 sign2 ...'(use space or line break to separate), and file whose name contains any one of them would be placed. As for why two output textboxes are placed here-well, it distinguishes upper and lower case of alphabets, if they are the same, it would be shown in left box, and right box on contrast. If more than one file meet the requirement, they would be display as well.