Create step-descent.py
Newton downhill method. Variable in func should be a tuple, y0 is the initial point. Default step is 1 (as a tuple input).
The function step is just a middle process of function stepde, so use stepde only, but step should be used together with stepde.
Function z is just an example to check the output of stepde.